Action Satoshis to int64

+++ wallet/serializer/list_actions.go
@@ -2,6 +2,7 @@ package serializer
 
 import (
 	"fmt"
+
 	"github.com/bsv-blockchain/go-sdk/chainhash"
 
 	"github.com/bsv-blockchain/go-sdk/util"
@@ -114,7 +115,7 @@ func SerializeListActionsResult(result *wallet.ListActionsResult) ([]byte, error
 	for _, action := range result.Actions {
 		// Serialize basic action fields
 		w.WriteBytes(action.Txid[:])
-		w.WriteVarInt(action.Satoshis)
+		w.WriteVarInt(uint64(action.Satoshis))
 
 		// Serialize status
 		switch action.Status {
@@ -193,7 +194,7 @@ func DeserializeListActionsResult(data []byte) (*wallet.ListActionsResult, error
 
 		// Deserialize basic action fields
 		copy(action.Txid[:], r.ReadBytes(chainhash.HashSize))
-		action.Satoshis = r.ReadVarInt()
+		action.Satoshis = int64(r.ReadVarInt())
 
 		// Deserialize status
 		status := r.ReadByte()
